betting-engine/requirements.txt
# Sports Betting Prediction Engine
# Python Dependencies

pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
ratelimit>=2.2.1
tabulate>=0.9.0
numpy>=1.24.0
